+++ buildSrc/src/main/java/co/anitrend/retrofit/graphql/buildSrc/plugin/components/AndroidPlugins.kt
@@ -18,7 +18,6 @@ private fun Project.applyModulePlugin() {
 internal fun Project.configurePlugins() {
     applyModulePlugin()
     plugins.apply("kotlin-android")
-    plugins.apply("kotlin-parcelize")
     if (isSampleModule())
         plugins.apply("kotlin-kapt")
 }